Revert "refactor dask converters (#1302)"
This reverts commit eafabc9b9ecbf0499d21e7ec62402ed3b8a871d8.
