+++ aeon/datatypes/_adapter/gluonts_to_pd_multiindex.py
@@ -1,7 +1,4 @@
-import pandas as pd
 from deprecated.sphinx import deprecated
-
-from aeon.datatypes import convert_to
 
 
 # TODO: remove in v0.8.0
@@ -32,6 +29,10 @@ def convert_gluonts_result_to_multiindex(gluonts_result):
     A MultiIndex DF mtype type compatible with aeon.
 
     """
+    import pandas as pd
+
+    from aeon.datatypes import convert_to
+
     instance_no = len(gluonts_result)
     global_ls = []
     per_instance_ls = []
