+++ aeon/datatypes/_adapter/tests/test_dask_pd.py
@@ -5,7 +5,7 @@ from copy import deepcopy
 import pandas as pd
 import pytest
 
-from aeon.utils.conversion.dask_converters import (
+from aeon.datatypes._adapter.dask_to_pd import (
     convert_dask_to_pandas,
     convert_pandas_to_dask,
 )
