+++ aeon/datatypes/_panel/_examples.py
@@ -81,7 +81,7 @@ example_dict[("nested_univ", "Panel", 0)] = X
 example_dict_lossy[("nested_univ", "Panel", 0)] = False
 
 if _check_soft_dependencies("dask", severity="none"):
-    from aeon.utils.conversion.dask_converters import convert_pandas_to_dask
+    from aeon.datatypes._adapter.dask_to_pd import convert_pandas_to_dask
 
     df_dask = convert_pandas_to_dask(
         example_dict[("pd-multiindex", "Panel", 0)], npartitions=1
@@ -151,7 +151,7 @@ example_dict[("nested_univ", "Panel", 1)] = X
 example_dict_lossy[("nested_univ", "Panel", 1)] = False
 
 if _check_soft_dependencies("dask", severity="none"):
-    from aeon.utils.conversion.dask_converters import convert_pandas_to_dask
+    from aeon.datatypes._adapter.dask_to_pd import convert_pandas_to_dask
 
     df_dask = convert_pandas_to_dask(
         example_dict[("pd-multiindex", "Panel", 1)], npartitions=1
@@ -216,7 +216,7 @@ example_dict[("nested_univ", "Panel", 2)] = X
 example_dict_lossy[("nested_univ", "Panel", 2)] = False
 
 if _check_soft_dependencies("dask", severity="none"):
-    from aeon.utils.conversion.dask_converters import convert_pandas_to_dask
+    from aeon.datatypes._adapter.dask_to_pd import convert_pandas_to_dask
 
     df_dask = convert_pandas_to_dask(
         example_dict[("pd-multiindex", "Panel", 2)], npartitions=1
